Add services StudentService and TestService
