+++ 02-test-student/src/main/java/ru/isakaev/Main.java
@@ -4,6 +4,7 @@ import org.springframework.context.annotation.AnnotationConfigApplicationContext
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import ru.isakaev.service.QuestionService;
+import ru.isakaev.service.TestService;
 
 @ComponentScan
 @Configuration
@@ -13,9 +14,7 @@ public class Main {
     public static void main(String[] args) {
 
         AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(Main.class);
-        QuestionService service = context.getBean(QuestionService.class);
-        service.loadQuestions();
+        TestService service = context.getBean(TestService.class);
         service.testStudent();
-
     }
 }
