+++ 02-test-student/src/main/java/ru/isakaev/service/StudentService.java
@@ -1,0 +1,8 @@
+package ru.isakaev.service;
+
+import ru.isakaev.model.Student;
+
+public interface StudentService {
+
+    Student getStudent();
+}
