+++ 02-test-student/src/main/java/ru/isakaev/service/TestService.java
@@ -1,0 +1,6 @@
+package ru.isakaev.service;
+
+public interface TestService {
+
+    void testStudent();
+}
